Migrate queries to the question form service
