+++ build.sbt
@@ -1,4 +1,4 @@
-import sbt.Project.projectToRef
+
 
 lazy val scalaV = "2.11.8"
